+++ Editor/Authoring/Deployment/LeaderboardDeploymentProvider.cs
@@ -1,5 +1,6 @@
 using System.Collections.ObjectModel;
 using Unity.Services.DeploymentApi.Editor;
+using Unity.Services.Leaderboards.Editor.Authoring.Commands;
 using UnityEditor;
 using UnityEngine;
 
@@ -14,11 +15,13 @@ namespace Unity.Services.Leaderboards.Editor.Authoring.Deployment
         public LeaderboardDeploymentProvider(
             DeployCommand deployCommand,
             OpenLeaderboardDashboardCommand openLeaderboardDashboardCommand,
-            ObservableCollection<IDeploymentItem> deploymentItems)
+            ObservableCollection<IDeploymentItem> deploymentItems,
+            ResetLeaderboardCommand resetLeaderboardCommand)
             : base(deploymentItems)
         {
             DeployCommand = deployCommand;
             Commands.Add(openLeaderboardDashboardCommand);
+            Commands.Add(resetLeaderboardCommand);
         }
     }
 }