+++ Editor/Shared/AssemblyInfo.cs
@@ -2,3 +2,6 @@ using System.Runtime.CompilerServices;
 
 [assembly: InternalsVisibleTo("Unity.Services.Leaderboards.Authoring.Client")]
 [assembly: InternalsVisibleTo("Unity.Services.Leaderboards.Editor.Authoring")]
+[assembly: InternalsVisibleTo("Unity.Services.Leaderboards.EditorTests")]
+[assembly: InternalsVisibleTo("DynamicProxyGenAssembly2")]
+