+++ Editor/Shared/Assets/AssetPostprocessorProxy.Gen.cs
@@ -1,9 +1,9 @@
 // WARNING: Auto generated code. Modifications will be lost!
-// Original source 'com.unity.services.shared' @0.0.11.
+// Original source 'com.unity.services.shared' @0.0.12.
 using System;
 using UnityEditor;
 
-namespace Unity.Services.Leaderboards.Assets
+namespace Unity.Services.Leaderboards.Editor.Shared.Assets
 {
     class AssetPostprocessorProxy : AssetPostprocessor
     {
@@ -15,6 +15,7 @@ namespace Unity.Services.Leaderboards.Assets
             remove => s_AllAssetsPostprocessed -= value;
         }
 
+        // ReSharper disable once UnusedMember.Local implicit usage
         static void OnPostprocessAllAssets(
             string[] importedAssetPaths,
             string[] deletedAssetPaths,